# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ColorGifts API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type BookStyle = (typeof BookStyle)[keyof typeof BookStyle];

export const BookStyle = {
  simple: "simple",
  cartoon: "cartoon",
  detailed: "detailed",
} as const;

export type BookStatus = (typeof BookStatus)[keyof typeof BookStatus];

export const BookStatus = {
  draft: "draft",
  generating: "generating",
  ready: "ready",
  ordered: "ordered",
} as const;

export interface Book {
  id: number;
  title: string;
  /** @nullable */
  subtitle?: string | null;
  /** @nullable */
  dedication?: string | null;
  style: BookStyle;
  status: BookStatus;
  /** @nullable */
  coverImagePath?: string | null;
  /** @nullable */
  shareToken?: string | null;
  pageCount: number;
  createdAt: string;
  updatedAt: string;
}

export type CreateBookBodyStyle =
  (typeof CreateBookBodyStyle)[keyof typeof CreateBookBodyStyle];

export const CreateBookBodyStyle = {
  simple: "simple",
  cartoon: "cartoon",
  detailed: "detailed",
} as const;

export interface CreateBookBody {
  title: string;
  /** @nullable */
  subtitle?: string | null;
  /** @nullable */
  dedication?: string | null;
  style: CreateBookBodyStyle;
}

export type UpdateBookBodyStyle =
  (typeof UpdateBookBodyStyle)[keyof typeof UpdateBookBodyStyle];

export const UpdateBookBodyStyle = {
  simple: "simple",
  cartoon: "cartoon",
  detailed: "detailed",
} as const;

export type UpdateBookBodyStatus =
  (typeof UpdateBookBodyStatus)[keyof typeof UpdateBookBodyStatus];

export const UpdateBookBodyStatus = {
  draft: "draft",
  generating: "generating",
  ready: "ready",
  ordered: "ordered",
} as const;

export interface UpdateBookBody {
  title?: string;
  /** @nullable */
  subtitle?: string | null;
  /** @nullable */
  dedication?: string | null;
  style?: UpdateBookBodyStyle;
  status?: UpdateBookBodyStatus;
  /** @nullable */
  coverImagePath?: string | null;
}

export type ColoringPageStatus =
  (typeof ColoringPageStatus)[keyof typeof ColoringPageStatus];

export const ColoringPageStatus = {
  pending: "pending",
  generating: "generating",
  ready: "ready",
  failed: "failed",
} as const;

export interface ColoringPage {
  id: number;
  bookId: number;
  photoId: number;
  originalImagePath: string;
  /** @nullable */
  coloringImagePath?: string | null;
  status: ColoringPageStatus;
  sortOrder: number;
  createdAt: string;
  updatedAt: string;
}

export interface Photo {
  id: number;
  bookId: number;
  objectPath: string;
  fileName: string;
  createdAt: string;
}

export interface CreatePhotoBody {
  bookId: number;
  objectPath: string;
  fileName: string;
}

export interface GenerateResult {
  message: string;
  pagesQueued: number;
}

export interface RequestUploadUrlBody {
  name: string;
  size: number;
  contentType: string;
}

export interface RequestUploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}