# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ColorGifts API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all coloring books
 */
export const ListBooksResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  subtitle: zod.string().nullish(),
  dedication: zod.string().nullish(),
  style: zod.enum(["simple", "cartoon", "detailed"]),
  status: zod.enum(["draft", "generating", "ready", "ordered"]),
  coverImagePath: zod.string().nullish(),
  shareToken: zod.string().nullish(),
  pageCount: zod.number(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const ListBooksResponse = zod.array(ListBooksResponseItem);

/**
 * @summary Create a new coloring book
 */
export const CreateBookBody = zod.object({
  title: zod.string(),
  subtitle: zod.string().nullish(),
  dedication: zod.string().nullish(),
  style: zod.enum(["simple", "cartoon", "detailed"]),
});

/**
 * @summary Get a book by ID
 */
export const GetBookParams = zod.object({
  id: zod.coerce.number(),
});

export const GetBookResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  subtitle: zod.string().nullish(),
  dedication: zod.string().nullish(),
  style: zod.enum(["simple", "cartoon", "detailed"]),
  status: zod.enum(["draft", "generating", "ready", "ordered"]),
  coverImagePath: zod.string().nullish(),
  shareToken: zod.string().nullish(),
  pageCount: zod.number(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Update a book
 */
export const UpdateBookParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateBookBody = zod.object({
  title: zod.string().optional(),
  subtitle: zod.string().nullish(),
  dedication: zod.string().nullish(),
  style: zod.enum(["simple", "cartoon", "detailed"]).optional(),
  status: zod.enum(["draft", "generating", "ready", "ordered"]).optional(),
  coverImagePath: zod.string().nullish(),
});

export const UpdateBookResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  subtitle: zod.string().nullish(),
  dedication: zod.string().nullish(),
  style: zod.enum(["simple", "cartoon", "detailed"]),
  status: zod.enum(["draft", "generating", "ready", "ordered"]),
  coverImagePath: zod.string().nullish(),
  shareToken: zod.string().nullish(),
  pageCount: zod.number(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Delete a book
 */
export const DeleteBookParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List all pages in a book
 */
export const ListBookPagesParams = zod.object({
  id: zod.coerce.number(),
});

export const ListBookPagesResponseItem = zod.object({
  id: zod.number(),
  bookId: zod.number(),
  photoId: zod.number(),
  originalImagePath: zod.string(),
  coloringImagePath: zod.string().nullish(),
  status: zod.enum(["pending", "generating", "ready", "failed"]),
  sortOrder: zod.number(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const ListBookPagesResponse = zod.array(ListBookPagesResponseItem);

/**
 * @summary Trigger AI generation for all uploaded photos in a book
 */
export const GenerateBookPagesParams = zod.object({
  id: zod.coerce.number(),
});

export const GenerateBookPagesResponse = zod.object({
  message: zod.string(),
  pagesQueued: zod.number(),
});

/**
 * @summary Get a coloring page by ID
 */
export const GetPageParams = zod.object({
  id: zod.coerce.number(),
});

export const GetPageResponse = zod.object({
  id: zod.number(),
  bookId: zod.number(),
  photoId: zod.number(),
  originalImagePath: zod.string(),
  coloringImagePath: zod.string().nullish(),
  status: zod.enum(["pending", "generating", "ready", "failed"]),
  sortOrder: zod.number(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Delete a coloring page
 */
export const DeletePageParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Regenerate a single coloring page
 */
export const RegeneratePageParams = zod.object({
  id: zod.coerce.number(),
});

export const RegeneratePageResponse = zod.object({
  id: zod.number(),
  bookId: zod.number(),
  photoId: zod.number(),
  originalImagePath: zod.string(),
  coloringImagePath: zod.string().nullish(),
  status: zod.enum(["pending", "generating", "ready", "failed"]),
  sortOrder: zod.number(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Request a presigned URL for file upload
 */
export const RequestUploadUrlBody = zod.object({
  name: zod.string(),
  size: zod.number(),
  contentType: zod.string(),
});

export const RequestUploadUrlResponse = zod.object({
  uploadURL: zod.string(),
  objectPath: zod.string(),
});

/**
 * @summary Serve a stored object
 */
export const GetObjectParams = zod.object({
  objectPath: zod.coerce.string(),
});

/**
 * @summary Register an uploaded photo for a book
 */
export const CreatePhotoBody = zod.object({
  bookId: zod.number(),
  objectPath: zod.string(),
  fileName: zod.string(),
});

/**
 * @summary Delete a photo
 */
export const DeletePhotoParams = zod.object({
  id: zod.coerce.number(),
});