# lib/api-zod/src/generated/types/book.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ColorGifts API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BookStatus } from "./bookStatus";
import type { BookStyle } from "./bookStyle";

export interface Book {
  id: number;
  title: string;
  /** @nullable */
  subtitle?: string | null;
  /** @nullable */
  dedication?: string | null;
  style: BookStyle;
  status: BookStatus;
  /** @nullable */
  coverImagePath?: string | null;
  /** @nullable */
  shareToken?: string | null;
  pageCount: number;
  createdAt: Date;
  updatedAt: Date;
}