# lib/api-zod/src/generated/types/coloringPage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ColorGifts API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ColoringPageStatus } from "./coloringPageStatus";

export interface ColoringPage {
  id: number;
  bookId: number;
  photoId: number;
  originalImagePath: string;
  /** @nullable */
  coloringImagePath?: string | null;
  status: ColoringPageStatus;
  sortOrder: number;
  createdAt: Date;
  updatedAt: Date;
}