# lib/api-zod/src/generated/types/createBookBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ColorGifts API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateBookBodyStyle } from "./createBookBodyStyle";

export interface CreateBookBody {
  title: string;
  /** @nullable */
  subtitle?: string | null;
  /** @nullable */
  dedication?: string | null;
  style: CreateBookBodyStyle;
}