# lib/api-zod/src/generated/types/createPhotoBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ColorGifts API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreatePhotoBody {
  bookId: number;
  objectPath: string;
  fileName: string;
}