# lib/api-zod/src/generated/types/photo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ColorGifts API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Photo {
  id: number;
  bookId: number;
  objectPath: string;
  fileName: string;
  createdAt: Date;
}