# lib/api-zod/src/generated/types/updateBookBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ColorGifts API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateBookBodyStatus } from "./updateBookBodyStatus";
import type { UpdateBookBodyStyle } from "./updateBookBodyStyle";

export interface UpdateBookBody {
  title?: string;
  /** @nullable */
  subtitle?: string | null;
  /** @nullable */
  dedication?: string | null;
  style?: UpdateBookBodyStyle;
  status?: UpdateBookBodyStatus;
  /** @nullable */
  coverImagePath?: string | null;
}